+++ javaweb/javaweb03/src/main/java/com/bobo/service/EmpService.java
@@ -8,7 +8,7 @@ public interface EmpService {
 
   public Emp selectOne(Emp emp);
 
-  public List<Emp> selectPage(int begin, int pageSize);
+  public List<Emp> selectPage(long begin, long pageSize);
 
   public Long selectCount();
